upgrade google storage adapter and fix client polling issues
