+++ packages/oc-gs-storage-adapter/index.js
@@ -6,7 +6,7 @@ const format = require('stringformat');
 const fs = require('fs-extra');
 const nodeDir = require('node-dir');
 const _ = require('lodash');
-const Storage = require('@google-cloud/storage');
+const { Storage } = require('@google-cloud/storage');
 
 const {
   getFileInfo,
@@ -21,12 +21,18 @@ module.exports = function(conf) {
     }
     return true;
   };
+
+  let client = undefined;
+
   const getClient = () => {
-    const client = Storage({
-      projectId: conf.projectId
-    });
+    if (!client) {
+      client = new Storage({
+        projectId: conf.projectId
+      });
+    }
     return client;
   };
+
   const bucketName = conf.bucket;
   const cache = new Cache({
     verbose: !!conf.verbosity,
@@ -71,6 +77,7 @@ module.exports = function(conf) {
 
     getFromGs((err, result) => {
       if (err) {
+        console.log('getFromGs: ', err, filePath);
         return callback({ code: err.code, msg: err.message });
       }
       cache.set('gs-file', filePath, result);
@@ -86,12 +93,14 @@ module.exports = function(conf) {
     }
     getFile(filePath, force, (err, file) => {
       if (err) {
+        console.log('getFile: ', err, filePath);
         return callback(err);
       }
 
       try {
         callback(null, JSON.parse(file));
       } catch (er) {
+        console.log('getJson: ', er, filePath);
         return callback({
           code: strings.errors.STORAGE.FILE_NOT_VALID_CODE,
           msg: format(strings.errors.STORAGE.FILE_NOT_VALID, filePath)
@@ -141,6 +150,9 @@ module.exports = function(conf) {
 
   const putDir = (dirInput, dirOutput, callback) => {
     nodeDir.paths(dirInput, (err, paths) => {
+      if (err) {
+        return callback(err, undefined);
+      }
       async.each(
         paths.files,
         (file, cb) => {
