+++ packages/oc-gs-storage-adapter/package.json
@@ -28,7 +28,7 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@google-cloud/storage": "^1.3.1",
+    "@google-cloud/storage": "2.5.0",
     "async": "^2.5.0",
     "fs-extra": "7.0.1",
     "lodash": "^4.17.4",
